Added RK3 for the homogeneous parts of the system (although including source terms of the electric field). Also added second order Strang splitting for the source terms of the form S(dt/2)H(dt)S(dt/2) where H(dt) is the RK3 homogeneous part.
Changed structure of the updates CAMReXmp::fluidSolverWENO() and CAMReXmp::MaxwellSolverDivFreeWENO() so that the first argument is the destination and the second is the source, just like MultiFab::Copy().
Fixed some bugs, especifically added boundary condition calculations
after the updates. Currently not supporting cylindrical geometry in
source terms, so changed CAMReXmp::sourceUpdate().

Also removed BC calculations in CAMReXmp::linearCombination() because it
assumes Phi_Type data, but now we have face centred data.

Currently using MultiFab::Copy() in the RK3 for FC data. This is because
for some reason the BC can't be calculated appropiately, leading to
undefined ghost cells at the domain boundaries. Need to check this in
the future. Ideally, MultiFab::Copy() should be avoided to avoid extra
computation.