Implicit Maxwell solver with constraint transport for the magnetic field using finite volume reconstruction at edges. Currently not working for constraint transport for the electric field, but there is a commented example code. Currently using finite differences for cell-centered Ex and Ey, and finite differences using finite volume reconstruction for Ez.
Added HYPRE in gnumakefile and in MaxwellFunctions.cpp. Could be useful
to solve the canonical form of the linear solver for face-centered
variables. Currently using AMReX's own linear solvers for the canonical
form, which is really only for cell-centered variables. Therefore
magnetic divergence errors could be seen at the top and bottom
boundaries of the reconnection challange.

Also added E_linearFunc() in utils, which only gives the reconstruction
of one field.

Commented out the refluxing of fluidFuctions.cpp if needed to be used
as source terms of face-centered electric fields.