It is unclear how the reflective boundary conditions should be implemented using HYPRE for the Harris test. Initially a reflective boundary condition was created in the previous commit but that created issues when using a high resolution (1024x512 cells). Then, the Dirichlet BC was considered when dealing with reflective, but it also causes some problems with high resolution tests. Therefore a new test was created, namely the Harris_full, which doubled the domain, created two current sheets and negated some values. This test reproduces the conducting wall boundary by using periodic boundary conditions in all directions. This is also done by Laguna et al.
Divergence cleaning variables source updates are commented out and there
are comments in CAMReXmp.cpp to comment the copy functions of the nodal
variables when using HDC (which is all cell-centered).

In the next commit, everything will tried to be more automatic, without
needing to comment and uncomment code when using different methods.