Changed the implicit midpoint solver for source terms based on Wang et al. 2020 such that the energies are updated based on the difference on kinetic energies. This is true because this source only allows energy transfer between the plasma and the electric field, and does not alter the pressure. This also avoids truncation errors that might lead to negative pressures.
The stiff Rosenbrock4 solver from boost/numeric/ is also changed so that
the updated variables are the momentum and the electric field. The
energy is updated in the same manner as the previously described.
However, it is found that this is significanly slower than the simple
midpoint method and does not give any further accuracy.

A newly created file sourceFunctionsExact.cpp has an exact source solver
based on https://github.com/ammarhakim/gkyl and especifically the
function gkyl/Updater/MomentSrcCommon.cpp. Again, no clear accuracy
improvements.