First, the code is made to run in 1D again because previously, the header files in MaxwellFunctions.cpp included HYPRE-related files, and these were intrisically multi-dimensional (at least it did not work in 1D). #if (AMREX_SPACEDIM >= 2) macros were used to make sure staggered grids related code and functionalities only appear in multiple dimensions. However, many functions, such as many RK functions, they are built based on the fact that FVTD Maxwell solvers were used. These type of functions were left but an abort call is made in CAMReXmp.cpp, especifically in read_params(), such that if multiple dimensions were used an FDTD or FVTD were also used, the code will abort.
Another big change is the inclusion of MaxwellSolverCN(), which is a
second-order Crank-Nicolson method for the Maxwell equations but using
cell-centered data (no FDTD). This will also work in 1D and AMR will
work more easily.

Next steps should include projection methods for the electric field so
that it works better for MaxwellSolverCN(). Maybe also delete all the
unnecessary code (such as those RK functions, and many more functions in
utils.cpp).