Added new files for the twof-fluid plasma equations with the generalized Ohm's law in the limit of small electron inertia and quasi-neutrality. Currently working with FORCE but not for large CFL nor for high resolution. Other flux methods like Rusanov or MUSTA are not working. Ideal Orszag-Tang is also not working, only 1D Brio-Wu.
Note that in this framework we are changing the charge to mass ratio,
not the ion Larmor ratiud (l_r=1).

Also added auxiliary functions (like FORCE) that works with any flux
function.

May need to change the underlying equations because something is not
working correctly. First try will be changing the ion energy equation to
a total energy equation.