IMEX for pressure sub-system correctly implemented. It now works for 1D Toro problem 1, although minor oscillations are present.
Used a Strang splitting for the pressure solve because it was found that
for CFL=0.8 the code crashes (but not for higher or lower CFL). This is
thought to be related to the fact that the advection is solved using a
second order time stepping, but only first order for the pressure. In
the future second order accurate IMEX-SSP(2,2,2) stiffly-accurate scheme
from Pareschi and Russo should be implemented to achieve second-order
temporal accuracy for the pressure.