Succesfully implemented IMEX-SSP(2,2,2) and IMEX-SSP(3,2,2) for pressure-based semi-implicit methods with a flattener algorithm that correctly removed the oscillations in Toro 1 problem. Also created a first-order in space and time flux function for that. This is idea is based on Boreschi and Pareschi (2021).
However, for other Toro problems the simulation still crashes. For Toro
test 3, a low CFL number needs to be used to run the simulation (CFL=0.4).
In the future, this might not be an issue since the fluid part is
coupled with the EM part, and such strong shocks might not be present,
and in multiple dimensions, the CFL is also reduced. This needs to be
explored further.