Corrected the main bug in the code, which is in the Rusanov flux calculation. In the speed estimate calculation, it was calculating the momentum, and it should be the velocity. This incorrectly calculated the dissipation.
Also corrected the way the pressure sub-system was solved. According to
the paper, the energy is updated after the Picard iteration was
finished, so the pressure is the variable that is tracked.

Also added the Dumbser version of the pressure-based semi-implicit
methods. This is based on staggered grids.

Currectly it produces oscillations in the full two-fluid system. It is
thought to be related that Strang splitting is used, and the source
terms should probably be in the IMEX-SSP scheme.

Also added TVD characteristic slope calculation, but it did not solve
the problems.