Added the first order Backward Euler implicit solver for the Maxwell equations. Added a sub-cycling algorithm for the electron fluxes with explicit flux calculations (no IMEX). This is used to test whether relaxing the sound speed constraint does produce reasonable solutions.
It was found that even for the second order Crank Nicolson implicit
solver for the Maxwell equations, with an second order implicit midpoint
method for the source terms, and fully explicit ion and electron fluxes,
the Brio and Wu test for l_r=0.001 still produces unphysical
oscillations. Note that this is without sub-cycling.

It was found that to reduce these oscillations, a first order Backward
Euler implicit solver for the Maxwell equations need to be used. A first
order implicit method for the source terms were tried and it does reduce
some oscillations, but the Maxwell part is more important. Therefore,
tests were carried using sub-cycling the electron fluxes, BE for the
Maxwell equations, and 2nd order implicit method for the source terms.
It was found that stepping over the full electron sound speed produces
oscillations. It was found that using a time step restricted by half or
a quarter of the electron sound speed relaxes the number of steps and
produces reasonable results (although solutions are more damped compared
to the fully explicit case). It is not clear whether stepping over the
electron sound speed is effective or not.

In the future, it might be important to remove these high frequency
oscillations found for l_r=0.001. It could be done by making the
approximation of neglecting displacement currect, which removes the time
derivative of the electric field giving an equation for the current
related to the magnetic field: J = curl(B)/mu_0